Displaying a sprite with OpenGL, finally

+++ assets/shaders/sprite.frag
@@ -1,6 +1,6 @@
 #version 330 core
 in vec2 f_texCoords;
-layout(location = 0) out vec4 color;
+layout(location = 0) out uvec4 color;
 
 uniform usampler2D glSpriteTexture;
